# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Enterprise API for EvalHub platform with role-based security",
        "title": "EvalHub API",
        "termsOfService": "https://evalhub.com/terms",
        "contact": {
            "name": "EvalHub API Support",
            "url": "https://evalhub.com/support",
            "email": "api-support@evalhub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session-based authentication cookie",
            "type": "apiKey",
            "name": "session_token",
            "in": "cookie"
        }
    },
    "paths": {
...
        "/jobs": {
            "get": {
                "security": [
                    {"BearerAuth": []},
                    {"SessionAuth": []}
                ],
                "description": "Retrieves a paginated list of jobs",
                "tags": ["Jobs"],
                "summary": "List jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "minimum": 1,
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100,
                        "default": 20,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by job type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "schema": {
                            "$ref": "#/definitions/JobsListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {"BearerAuth": []},
                    {"SessionAuth": []}
                ],
                "description": "Creates a new job posting",
                "tags": ["Jobs"],
                "summary": "Create a new job",
                "parameters": [
                    {
                        "description": "Job creation data",
                        "name": "createJobRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Job created successfully",
                        "schema": {
                            "$ref": "#/definitions/JobResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "error": {
                    "type": "string",
                    "example": "An error occurred"
                },
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_123456789"
                }
            }
        },
...
        "UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "enum": ["expert", "evaluator", "admin"],
                    "example": "expert"
                }
            }
        },
...
        "CreateJobRequest": {
            "type": "object",
            "required": [
                "title",
                "description",
                "company",
                "location",
                "job_type",
                "salary_range"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "example": "Senior Go Developer"
                },
                "description": {
                    "type": "string",
                    "minLength": 1,
                    "example": "We are looking for an experienced Go developer to join our team..."
                },
                "company": {
                    "type": "string",
                    "example": "TechCorp Inc."
                },
                "location": {
                    "type": "string",
                    "example": "New York, NY"
                },
                "job_type": {
                    "type": "string",
                    "enum": ["full-time", "part-time", "contract", "freelance", "internship"],
                    "example": "full-time"
                },
                "salary_range": {
                    "type": "string",
                    "example": "$100,000 - $150,000"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["5+ years Go experience", "Experience with microservices", "Knowledge of Docker"]
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["Health insurance", "401k matching", "Remote work"]
                },
                "remote_ok": {
                    "type": "boolean",
                    "default": false,
                    "example": true
                },
                "is_featured": {
                    "type": "boolean",
                    "default": false,
                    "example": false
                }
            }
        },
        "JobResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Senior Go Developer"
                },
                "description": {
                    "type": "string",
                    "example": "We are looking for an experienced Go developer..."
                },
                "company": {
                    "type": "string",
                    "example": "TechCorp Inc."
                },
                "location": {
                    "type": "string",
                    "example": "New York, NY"
                },
                "job_type": {
                    "type": "string",
                    "enum": ["full-time", "part-time", "contract", "freelance", "internship"],
                    "example": "full-time"
                },
                "salary_range": {
                    "type": "string",
                    "example": "$100,000 - $150,000"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["5+ years Go experience", "Experience with microservices"]
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": ["Health insurance", "401k matching"]
                },
                "remote_ok": {
                    "type": "boolean",
                    "example": true
                },
                "is_featured": {
                    "type": "boolean",
                    "example": false
                },
                "employer": {
                    "$ref": "#/definitions/UserInfo"
                },
                "applications_count": {
                    "type": "integer",
                    "example": 15
                },
                "views_count": {
                    "type": "integer",
                    "example": 234
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T12:00:00Z"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T12:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-02-01T12:00:00Z"
                }
            }
        },
        "JobsListResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/JobResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/PaginationMeta"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_123456789"
                }
            }
        },
        "PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "format": "int64",
                    "example": 20
                },
                "total_pages": {
                    "type": "integer",
                    "format": "int64",
                    "example": 5
                },
                "total_count": {
                    "type": "integer",
                    "format": "int64",
                    "example": 95
                },
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_prev": {
                    "type": "boolean",
                    "example": false
                }
            }
        }
    },
...
}